Change MediaHandler to SessionDescriptionHandler   - A lot of functionality was merged into the SessionDescriptionHandler     - Allow description modifiers to be used instead of forcing a full handler to be written     - Remove dependencies on enviornment and WebRTC     - Remove any sort of notion of "supported"     - Remove actual handling of media, this should be done application level     - Session description check just checks that the content type matches, leave the message parsing to the session layer   - Move away from using enviornment, instead just allow the user to provide their own SessionDescriptionHandler   - Remove browser prefixes, browsers no longer require prefixed WebRTC commands   - Remove MediaStream Manager   - Remove WebRTC.js to make SIP.js more agnostic to the underlying media
TODO:
  - Implement SessionDescriptionHandler.close()
  - Events
  - Comments
  - Tests (150 failing at time of commit)
  - Determine if promisifiy is still needed for a lot of description handling code
  - SDP modifying options on the UA need to be read by the descriptionHandler and coded up.
    - Make sure there are no sdp modifications happening in Session.js
  - UA
  - Session
  - Documentation
