+++ src/SIP.js
@@ -34,12 +34,11 @@ require('./Transactions')(SIP);
 require('./Dialogs')(SIP);
 require('./RequestSender')(SIP);
 require('./RegisterContext')(SIP);
-SIP.MediaHandler = require('./MediaHandler')(SIP.EventEmitter);
+SIP.SessionDescriptionHandler = require('./SessionDescriptionHandler')(SIP.EventEmitter);
 require('./ClientContext')(SIP);
 require('./ServerContext')(SIP);
 require('./Session')(SIP, environment);
 require('./Subscription')(SIP);
-SIP.WebRTC = require('./WebRTC')(SIP, environment);
 require('./UA')(SIP, environment);
 require('./SanityCheck')(SIP);
 SIP.DigestAuthentication = require('./DigestAuthentication')(SIP.Utils);
