--- src/WebRTC/MediaHandler.js
@@ -1,618 +1,0 @@
-"use strict";
-/**
- * @fileoverview MediaHandler
- */
-
-/* MediaHandler
- * @class PeerConnection helper Class.
- * @param {SIP.Session} session
- * @param {Object} [options]
- * @param {SIP.WebRTC.MediaStreamManager} [options.mediaStreamManager]
- *        The MediaStreamManager to acquire/release streams from/to.
- *        If not provided, a default MediaStreamManager will be used.
- */
-module.exports = function (SIP) {
-
-var MediaHandler = function(session, options) {
-  options = options || {};
-
-  this.logger = session.ua.getLogger('sip.invitecontext.mediahandler', session.id);
-  this.session = session;
-  this.localMedia = null;
-  this.ready = true;
-  this.mediaStreamManager = options.mediaStreamManager || new SIP.WebRTC.MediaStreamManager(this.logger);
-  this.audioMuted = false;
-  this.videoMuted = false;
-  this.local_hold = false;
-  this.remote_hold = false;
-
-  // old init() from here on
-  var servers = this.prepareIceServers(options.stunServers, options.turnServers);
-  this.RTCConstraints = options.RTCConstraints || {};
-
-  this.initPeerConnection(servers);
-
-  function selfEmit(mh, event) {
-    if (mh.mediaStreamManager.on) {
-      mh.mediaStreamManager.on(event, function () {
-        mh.emit.apply(mh, [event].concat(Array.prototype.slice.call(arguments)));
-      });
-    }
-  }
-
-  selfEmit(this, 'userMediaRequest');
-  selfEmit(this, 'userMedia');
-  selfEmit(this, 'userMediaFailed');
-};
-
-MediaHandler.defaultFactory = function defaultFactory (session, options) {
-  return new MediaHandler(session, options);
-};
-MediaHandler.defaultFactory.isSupported = function () {
-  return SIP.WebRTC.isSupported();
-};
-
-MediaHandler.prototype = Object.create(SIP.MediaHandler.prototype, {
-// Functions the session can use
-  isReady: {writable: true, value: function isReady () {
-    return this.ready;
-  }},
-
-  close: {writable: true, value: function close () {
-    this.logger.log('closing PeerConnection');
-    this._remoteStreams = [];
-    // have to check signalingState since this.close() gets called multiple times
-    // TODO figure out why that happens
-    if(this.peerConnection && this.peerConnection.signalingState !== 'closed') {
-      this.peerConnection.close();
-
-      if(this.localMedia) {
-        this.mediaStreamManager.release(this.localMedia);
-      }
-    }
-  }},
-
-  /**
-   * @param {SIP.WebRTC.MediaStream | (getUserMedia constraints)} [mediaHint]
-   *        the MediaStream (or the constraints describing it) to be used for the session
-   */
-  getDescription: {writable: true, value: function getDescription (mediaHint) {
-    var self = this;
-    var acquire = self.mediaStreamManager.acquire;
-    if (acquire.length > 1) {
-      acquire = SIP.Utils.promisify(this.mediaStreamManager, 'acquire', true);
-    }
-    mediaHint = mediaHint || {};
-    if (mediaHint.dataChannel === true) {
-      mediaHint.dataChannel = {};
-    }
-    this.mediaHint = mediaHint;
-
-    /*
-     * 1. acquire streams (skip if MediaStreams passed in)
-     * 2. addStreams
-     * 3. createOffer/createAnswer
-     */
-
-    var streamPromise;
-    if (self.localMedia) {
-      self.logger.log('already have local media');
-      streamPromise = SIP.Utils.Promise.resolve(self.localMedia);
-    }
-    else {
-      self.logger.log('acquiring local media');
-      streamPromise = acquire.call(self.mediaStreamManager, mediaHint)
-        .then(function acquireSucceeded(streams) {
-          self.logger.log('acquired local media streams');
-          self.localMedia = streams;
-          self.session.connecting();
-          return streams;
-        }, function acquireFailed(err) {
-          self.logger.error('unable to acquire streams');
-          self.logger.error(err);
-          self.session.connecting();
-          throw err;
-        })
-        .then(this.addStreams.bind(this))
-      ;
-    }
-
-    return streamPromise
-      .then(function streamAdditionSucceeded() {
-        if (self.hasOffer('remote')) {
-          self.peerConnection.ondatachannel = function (evt) {
-            self.dataChannel = evt.channel;
-            self.emit('dataChannel', self.dataChannel);
-          };
-        } else if (mediaHint.dataChannel &&
-                   self.peerConnection.createDataChannel) {
-          self.dataChannel = self.peerConnection.createDataChannel(
-            'sipjs',
-            mediaHint.dataChannel
-          );
-          self.emit('dataChannel', self.dataChannel);
-        }
-
-        self.render();
-        return self.createOfferOrAnswer(self.RTCConstraints);
-      })
-      .then(function(sdp) {
-
-        if (self.local_hold) {
-          // Don't receive media
-          // TODO - This will break for media streams with different directions.
-          if (!(/a=(sendrecv|sendonly|recvonly|inactive)/).test(sdp)) {
-            sdp = sdp.replace(/(m=[^\r]*\r\n)/g, '$1a=sendonly\r\n');
-          } else {
-            sdp = sdp.replace(/a=sendrecv\r\n/g, 'a=sendonly\r\n');
-            sdp = sdp.replace(/a=recvonly\r\n/g, 'a=inactive\r\n');
-          }
-        }
-
-        return {
-          body: sdp,
-          contentType: 'application/sdp'
-        };
-      })
-    ;
-  }},
-
-  /**
-   * Check if a SIP message contains a session description.
-   * @param {SIP.SIPMessage} message
-   * @returns {boolean}
-   */
-  hasDescription: {writeable: true, value: function hasDescription (message) {
-    return message.getHeader('Content-Type') === 'application/sdp' && !!message.body;
-  }},
-
-  /**
-   * Set the session description contained in a SIP message.
-   * @param {SIP.SIPMessage} message
-   * @returns {Promise}
-   */
-  setDescription: {writable: true, value: function setDescription (message) {
-    var self = this;
-    var sdp = message.body;
-
-    this.remote_hold = /a=(sendonly|inactive)/.test(sdp);
-
-
-    var rawDescription = {
-      type: this.hasOffer('local') ? 'answer' : 'offer',
-      sdp: sdp
-    };
-
-    this.emit('setDescription', rawDescription);
-
-    var description = new SIP.WebRTC.RTCSessionDescription(rawDescription);
-    return SIP.Utils.promisify(this.peerConnection, 'setRemoteDescription')(description)
-      .catch(function setRemoteDescriptionError(e) {
-        self.emit('peerConnection-setRemoteDescriptionFailed', e);
-        throw e;
-      });
-  }},
-
-  /**
-   * If the Session associated with this MediaHandler were to be referred,
-   * what mediaHint should be provided to the UA's invite method?
-   */
-  getReferMedia: {writable: true, value: function getReferMedia () {
-    function hasTracks (trackGetter, stream) {
-      return stream[trackGetter]().length > 0;
-    }
-
-    function bothHaveTracks (trackGetter) {
-      /* jshint validthis:true */
-      return this.getLocalStreams().some(hasTracks.bind(null, trackGetter)) &&
-             this.getRemoteStreams().some(hasTracks.bind(null, trackGetter));
-    }
-
-    return {
-      constraints: {
-        audio: bothHaveTracks.call(this, 'getAudioTracks'),
-        video: bothHaveTracks.call(this, 'getVideoTracks')
-      }
-    };
-  }},
-
-  updateIceServers: {writeable:true, value: function (options) {
-    var servers = this.prepareIceServers(options.stunServers, options.turnServers);
-    this.RTCConstraints = options.RTCConstraints || this.RTCConstraints;
-
-    this.initPeerConnection(servers);
-
-    /* once updateIce is implemented correctly, this is better than above
-    //no op if browser does not support this
-    if (!this.peerConnection.updateIce) {
-      return;
-    }
-
-    this.peerConnection.updateIce({'iceServers': servers}, this.RTCConstraints);
-    */
-  }},
-
-// Functions the session can use, but only because it's convenient for the application
-  isMuted: {writable: true, value: function isMuted () {
-    return {
-      audio: this.audioMuted,
-      video: this.videoMuted
-    };
-  }},
-
-  mute: {writable: true, value: function mute (options) {
-    if (this.getLocalStreams().length === 0) {
-      return;
-    }
-
-    options = options || {
-      audio: this.getLocalStreams()[0].getAudioTracks().length > 0,
-      video: this.getLocalStreams()[0].getVideoTracks().length > 0
-    };
-
-    var audioMuted = false,
-        videoMuted = false;
-
-    if (options.audio && !this.audioMuted) {
-      audioMuted = true;
-      this.audioMuted = true;
-      this.toggleMuteAudio(true);
-    }
-
-    if (options.video && !this.videoMuted) {
-      videoMuted = true;
-      this.videoMuted = true;
-      this.toggleMuteVideo(true);
-    }
-
-    //REVISIT
-    if (audioMuted || videoMuted) {
-      return {
-        audio: audioMuted,
-        video: videoMuted
-      };
-      /*this.session.onmute({
-        audio: audioMuted,
-        video: videoMuted
-      });*/
-    }
-  }},
-
-  unmute: {writable: true, value: function unmute (options) {
-    if (this.getLocalStreams().length === 0) {
-      return;
-    }
-
-    options = options || {
-      audio: this.getLocalStreams()[0].getAudioTracks().length > 0,
-      video: this.getLocalStreams()[0].getVideoTracks().length > 0
-    };
-
-    var audioUnMuted = false,
-        videoUnMuted = false;
-
-    if (options.audio && this.audioMuted) {
-      audioUnMuted = true;
-      this.audioMuted = false;
-      this.toggleMuteAudio(false);
-    }
-
-    if (options.video && this.videoMuted) {
-      videoUnMuted = true;
-      this.videoMuted = false;
-      this.toggleMuteVideo(false);
-    }
-
-    //REVISIT
-    if (audioUnMuted || videoUnMuted) {
-      return {
-        audio: audioUnMuted,
-        video: videoUnMuted
-      };
-      /*this.session.onunmute({
-        audio: audioUnMuted,
-        video: videoUnMuted
-      });*/
-    }
-  }},
-
-  hold: {writable: true, value: function hold () {
-    this.local_hold = true;
-    this.toggleMuteAudio(true);
-    this.toggleMuteVideo(true);
-  }},
-
-  unhold: {writable: true, value: function unhold () {
-    this.local_hold = false;
-
-    if (!this.audioMuted) {
-      this.toggleMuteAudio(false);
-    }
-
-    if (!this.videoMuted) {
-      this.toggleMuteVideo(false);
-    }
-  }},
-
-// Functions the application can use, but not the session
-  getLocalStreams: {writable: true, value: function getLocalStreams () {
-    var pc = this.peerConnection;
-    if (pc && pc.signalingState === 'closed') {
-      this.logger.warn('peerConnection is closed, getLocalStreams returning []');
-      return [];
-    }
-    return (pc.getLocalStreams && pc.getLocalStreams()) ||
-      pc.localStreams || [];
-  }},
-
-  getRemoteStreams: {writable: true, value: function getRemoteStreams () {
-    var pc = this.peerConnection;
-    if (pc && pc.signalingState === 'closed') {
-      this.logger.warn('peerConnection is closed, getRemoteStreams returning this._remoteStreams');
-      return this._remoteStreams;
-    }
-    return(pc.getRemoteStreams && pc.getRemoteStreams()) ||
-      pc.remoteStreams || [];
-  }},
-
-  render: {writable: true, value: function render (renderHint) {
-    renderHint = renderHint || (this.mediaHint && this.mediaHint.render);
-    if (!renderHint) {
-      return false;
-    }
-    var streamGetters = {
-      local: 'getLocalStreams',
-      remote: 'getRemoteStreams'
-    };
-    Object.keys(streamGetters).forEach(function (loc) {
-      var streamGetter = streamGetters[loc];
-      var streams = this[streamGetter]();
-      SIP.WebRTC.MediaStreamManager.render(streams, renderHint[loc]);
-    }.bind(this));
-  }},
-
-// Internal functions
-  hasOffer: {writable: true, value: function hasOffer (where) {
-    var offerState = 'have-' + where + '-offer';
-    return this.peerConnection.signalingState === offerState;
-    // TODO consider signalingStates with 'pranswer'?
-  }},
-
-  prepareIceServers: {writable: true, value: function prepareIceServers (stunServers, turnServers) {
-    var servers = [],
-      config = this.session.ua.configuration;
-
-    stunServers = stunServers || config.stunServers;
-    turnServers = turnServers || config.turnServers;
-
-    [].concat(stunServers).forEach(function (server) {
-      servers.push({'urls': server});
-    });
-
-    [].concat(turnServers).forEach(function (server) {
-      var turnServer = {'urls': server.urls};
-      if (server.username) {
-        turnServer.username = server.username;
-      }
-      if (server.password) {
-        turnServer.credential = server.password;
-      }
-      servers.push(turnServer);
-    });
-
-    return servers;
-  }},
-
-  initPeerConnection: {writable: true, value: function initPeerConnection(servers) {
-    var self = this,
-      config = this.session.ua.configuration;
-
-    this.onIceCompleted = SIP.Utils.defer();
-    this.onIceCompleted.promise.then(function(pc) {
-      self.emit('iceGatheringComplete', pc);
-      if (self.iceCheckingTimer) {
-        SIP.Timers.clearTimeout(self.iceCheckingTimer);
-        self.iceCheckingTimer = null;
-      }
-    });
-
-    if (this.peerConnection) {
-      this.peerConnection.close();
-    }
-
-    var connConfig = {
-      iceServers: servers
-    };
-
-    if (config.rtcpMuxPolicy) {
-      connConfig.rtcpMuxPolicy = config.rtcpMuxPolicy;
-    }
-
-    this.peerConnection = new SIP.WebRTC.RTCPeerConnection(connConfig);
-
-    // Firefox (35.0.1) sometimes throws on calls to peerConnection.getRemoteStreams
-    // even if peerConnection.onaddstream was just called. In order to make
-    // MediaHandler.prototype.getRemoteStreams work, keep track of them manually
-    this._remoteStreams = [];
-
-    this.peerConnection.onaddstream = function(e) {
-      self.logger.log('stream added: '+ e.stream.id);
-      self._remoteStreams.push(e.stream);
-      self.render();
-      self.emit('addStream', e);
-    };
-
-    this.peerConnection.onremovestream = function(e) {
-      self.logger.log('stream removed: '+ e.stream.id);
-    };
-
-    this.startIceCheckingTimer = function () {
-      if (!self.iceCheckingTimer) {
-        self.iceCheckingTimer = SIP.Timers.setTimeout(function() {
-          self.logger.log('RTCIceChecking Timeout Triggered after '+config.iceCheckingTimeout+' milliseconds');
-          self.onIceCompleted.resolve(this);
-        }.bind(this.peerConnection), config.iceCheckingTimeout);
-      }
-    };
-
-    this.peerConnection.onicecandidate = function(e) {
-      self.emit('iceCandidate', e);
-      if (e.candidate) {
-        self.logger.log('ICE candidate received: '+ (e.candidate.candidate === null ? null : e.candidate.candidate.trim()));
-        self.startIceCheckingTimer();
-      } else {
-        self.onIceCompleted.resolve(this);
-      }
-    };
-
-    this.peerConnection.onicegatheringstatechange = function () {
-      self.logger.log('RTCIceGatheringState changed: ' + this.iceGatheringState);
-      if (this.iceGatheringState === 'gathering') {
-        self.emit('iceGathering', this);
-      }
-      if (this.iceGatheringState === 'complete') {
-        self.onIceCompleted.resolve(this);
-      }
-    };
-
-    this.peerConnection.oniceconnectionstatechange = function() {  //need e for commented out case
-      var stateEvent;
-
-      if (this.iceConnectionState === 'checking') {
-        self.startIceCheckingTimer();
-      }
-
-      switch (this.iceConnectionState) {
-      case 'new':
-        stateEvent = 'iceConnection';
-        break;
-      case 'checking':
-        stateEvent = 'iceConnectionChecking';
-        break;
-      case 'connected':
-        stateEvent = 'iceConnectionConnected';
-        break;
-      case 'completed':
-        stateEvent = 'iceConnectionCompleted';
-        break;
-      case 'failed':
-        stateEvent = 'iceConnectionFailed';
-        break;
-      case 'disconnected':
-        stateEvent = 'iceConnectionDisconnected';
-        break;
-      case 'closed':
-        stateEvent = 'iceConnectionClosed';
-        break;
-      default:
-        self.logger.warn('Unknown iceConnection state:', this.iceConnectionState);
-        return;
-      }
-      self.emit(stateEvent, this);
-
-      //Bria state changes are always connected -> disconnected -> connected on accept, so session gets terminated
-      //normal calls switch from failed to connected in some cases, so checking for failed and terminated
-      /*if (this.iceConnectionState === 'failed') {
-        self.session.terminate({
-        cause: SIP.C.causes.RTP_TIMEOUT,
-        status_code: 200,
-        reason_phrase: SIP.C.causes.RTP_TIMEOUT
-      });
-      } else if (e.currentTarget.iceGatheringState === 'complete' && this.iceConnectionState !== 'closed') {
-      self.onIceCompleted(this);
-      }*/
-    };
-
-    this.peerConnection.onstatechange = function() {
-      self.logger.log('PeerConnection state changed to "'+ this.readyState +'"');
-    };
-  }},
-
-  createOfferOrAnswer: {writable: true, value: function createOfferOrAnswer (constraints) {
-    var self = this;
-    var methodName;
-    var pc = self.peerConnection;
-
-    self.ready = false;
-    methodName = self.hasOffer('remote') ? 'createAnswer' : 'createOffer';
-
-    return SIP.Utils.promisify(pc, methodName, true)(constraints)
-      .catch(function methodError(e) {
-        self.emit('peerConnection-' + methodName + 'Failed', e);
-        throw e;
-      })
-      .then(SIP.Utils.promisify(pc, 'setLocalDescription'))
-      .catch(function localDescError(e) {
-        self.emit('peerConnection-selLocalDescriptionFailed', e);
-        throw e;
-      })
-      .then(function onSetLocalDescriptionSuccess() {
-        var deferred = SIP.Utils.defer();
-        if (pc.iceGatheringState === 'complete' && (pc.iceConnectionState === 'connected' || pc.iceConnectionState === 'completed')) {
-          deferred.resolve();
-        } else {
-          self.onIceCompleted.promise.then(deferred.resolve);
-        }
-        return deferred.promise;
-      })
-      .then(function readySuccess () {
-        var sdp = pc.localDescription.sdp;
-
-
-        var sdpWrapper = {
-          type: methodName === 'createOffer' ? 'offer' : 'answer',
-          sdp: sdp
-        };
-
-        self.emit('getDescription', sdpWrapper);
-
-        if (self.session.ua.configuration.hackStripTcp) {
-          sdpWrapper.sdp = sdpWrapper.sdp.replace(/^a=candidate:\d+ \d+ tcp .*?\r\n/img, "");
-        }
-
-        self.ready = true;
-        return sdpWrapper.sdp;
-      })
-      .catch(function createOfferAnswerError (e) {
-        self.logger.error(e);
-        self.ready = true;
-        throw new SIP.Exceptions.GetDescriptionError(e);
-      })
-    ;
-  }},
-
-  addStreams: {writable: true, value: function addStreams (streams) {
-    try {
-      streams = [].concat(streams);
-      streams.forEach(function (stream) {
-        this.peerConnection.addStream(stream);
-      }, this);
-    } catch(e) {
-      this.logger.error('error adding stream');
-      this.logger.error(e);
-      return SIP.Utils.Promise.reject(e);
-    }
-
-    return SIP.Utils.Promise.resolve();
-  }},
-
-  toggleMuteHelper: {writable: true, value: function toggleMuteHelper (trackGetter, mute) {
-    this.getLocalStreams().forEach(function (stream) {
-      stream[trackGetter]().forEach(function (track) {
-        track.enabled = !mute;
-      });
-    });
-  }},
-
-  toggleMuteAudio: {writable: true, value: function toggleMuteAudio (mute) {
-    this.toggleMuteHelper('getAudioTracks', mute);
-  }},
-
-  toggleMuteVideo: {writable: true, value: function toggleMuteVideo (mute) {
-    this.toggleMuteHelper('getVideoTracks', mute);
-  }}
-});
-
-// Return since it will be assigned to a variable.
-return MediaHandler;
-};
