--- src/WebRTC/MediaStreamManager.js
@@ -1,157 +1,0 @@
-"use strict";
-/**
- * @fileoverview MediaStreamManager
- */
-
-/* MediaStreamManager
- * @class Manages the acquisition and release of MediaStreams.
- * @param {mediaHint} [defaultMediaHint] The mediaHint to use if none is provided to acquire()
- */
-module.exports = function (SIP, environment) {
-
-// Default MediaStreamManager provides single-use streams created with getUserMedia
-var MediaStreamManager = function MediaStreamManager (logger, defaultMediaHint) {
-  if (!SIP.WebRTC.isSupported()) {
-    throw new SIP.Exceptions.NotSupportedError('Media not supported');
-  }
-
-  this.mediaHint = defaultMediaHint || {
-    constraints: {audio: true, video: true}
-  };
-
-  // map of streams to acquisition manner:
-  // true -> passed in as mediaHint.stream
-  // false -> getUserMedia
-  this.acquisitions = {};
-};
-MediaStreamManager.streamId = function (stream) {
-  return stream.getAudioTracks().concat(stream.getVideoTracks())
-    .map(function trackId (track) {
-      return track.id;
-    })
-    .join('');
-};
-
-/**
- * @param {(Array of) MediaStream} streams - The streams to render
- *
- * @param {(Array of) HTMLMediaElement} elements
- *        - The <audio>/<video> element(s) that should render the streams
- *
- * Each stream in streams renders to the corresponding element in elements,
- * wrapping around elements if needed.
- */
-MediaStreamManager.render = function render (streams, elements) {
-  if (!elements) {
-    return false;
-  }
-  if (Array.isArray(elements) && !elements.length) {
-    throw new TypeError('elements must not be empty');
-  }
-
-  function attachMediaStream(element, stream) {
-    element.srcObject = stream;
-  }
-
-  function ensureMediaPlaying (mediaElement) {
-    var interval = 100;
-    mediaElement.ensurePlayingIntervalId = SIP.Timers.setInterval(function () {
-      if (mediaElement.paused && mediaElement.srcObject) {
-        mediaElement.play();
-      }
-      else {
-        SIP.Timers.clearInterval(mediaElement.ensurePlayingIntervalId);
-      }
-    }, interval);
-  }
-
-  function attachAndPlay (elements, stream, index) {
-    var element = elements[index % elements.length];
-    if (typeof element === 'function') {
-      element = element();
-    }
-    (environment.attachMediaStream || attachMediaStream)(element, stream);
-    ensureMediaPlaying(element);
-  }
-
-  // [].concat "casts" `elements` into an array
-  // so forEach works even if `elements` was a single element
-  elements = [].concat(elements);
-  [].concat(streams).forEach(attachAndPlay.bind(null, elements));
-};
-
-MediaStreamManager.prototype = Object.create(SIP.EventEmitter.prototype, {
-  'acquire': {writable: true, value: function acquire (mediaHint) {
-    mediaHint = Object.keys(mediaHint || {}).length ? mediaHint : this.mediaHint;
-
-    var saveSuccess = function (isHintStream, streams) {
-      streams = [].concat(streams);
-      streams.forEach(function (stream) {
-        var streamId = MediaStreamManager.streamId(stream);
-        this.acquisitions[streamId] = !!isHintStream;
-      }, this);
-      return SIP.Utils.Promise.resolve(streams);
-    }.bind(this);
-
-    if (mediaHint.stream) {
-      return saveSuccess(true, mediaHint.stream);
-    } else {
-      // Fallback to audio/video enabled if no mediaHint can be found.
-      var constraints = mediaHint.constraints ||
-        (this.mediaHint && this.mediaHint.constraints) ||
-        {audio: true, video: true};
-
-      var deferred = SIP.Utils.defer();
-
-      /*
-       * Make the call asynchronous, so that ICCs have a chance
-       * to define callbacks to `userMediaRequest`
-       */
-      SIP.Timers.setTimeout(function () {
-        this.emit('userMediaRequest', constraints);
-
-        var emitThenCall = function (eventName, callback) {
-          var callbackArgs = Array.prototype.slice.call(arguments, 2);
-          // Emit with all of the arguments from the real callback.
-          var newArgs = [eventName].concat(callbackArgs);
-
-          this.emit.apply(this, newArgs);
-
-          return callback.apply(null, callbackArgs);
-        }.bind(this);
-
-        if (constraints.audio || constraints.video) {
-          deferred.resolve(
-            SIP.WebRTC.getUserMedia(constraints)
-            .then(
-              emitThenCall.bind(this, 'userMedia', saveSuccess.bind(null, false)),
-              emitThenCall.bind(this, 'userMediaFailed', function(e){throw e;})
-            )
-          );
-        } else {
-          // Local streams were explicitly excluded.
-          deferred.resolve([]);
-        }
-      }.bind(this), 0);
-
-      return deferred.promise;
-    }
-  }},
-
-  'release': {writable: true, value: function release (streams) {
-    streams = [].concat(streams);
-    streams.forEach(function (stream) {
-      var streamId = MediaStreamManager.streamId(stream);
-      if (this.acquisitions[streamId] === false) {
-        stream.getTracks().forEach(function (track) {
-          track.stop();
-        });
-      }
-      delete this.acquisitions[streamId];
-    }, this);
-  }},
-});
-
-// Return since it will be assigned to a variable.
-return MediaStreamManager;
-};
